+++ CHANGELOG.md
@@ -6,3 +6,6 @@
 - *Removed extra editor for when the explorer is not visible; updated the code to handle this in only one editor*
 - merged all explorer configs into one tag; also added new config default explorer width
 - *added new dependency `react-terminal`*
+- Add Terminal and terminal config
+- Added [nushell](https://github.com/nushell/nushell.git) as the default shell
+- moved languageExtensionMap from App.jsx to config